SLEMA-97 Moved SymptomsButtomBar to another file
